+++ test/integration/images_test.go
@@ -39,6 +39,23 @@ func TestImagesIntegration(t *testing.T) {
 		}
 	})
 
+	t.Run("get images by instance type", func(t *testing.T) {
+		images, err := client.Images.GetImagesByInstanceType(ctx, "8B300.240V")
+		if err != nil {
+			t.Errorf("failed to get images by instance type: %v", err)
+		}
+		t.Logf("Found %d images for instance type 8B300.240V", len(images))
+
+		if len(images) > 0 {
+			for i, img := range images {
+				if i < 3 {
+					t.Logf("Image: %s (%s) - Default: %v, Category: %s",
+						img.Name, img.ImageType, img.IsDefault, img.Category)
+				}
+			}
+		}
+	})
+
 	t.Run("get cluster images", func(t *testing.T) {
 		clusterImages, err := client.Images.GetClusterImages(ctx)
 		if err != nil {